"expiry window" cannot be done the same way in sigv4 as sigv2. temporarily disabling for now.  will revisit later.
Sigv4 X-Amz-Expires operates on a seconds until expiration basis, as
opposed to an explicit expiration timestamp in sigv2. Emulating the
later by manipulating either the X-Amz-Expires or X-Amz-Date will not
automatically produce stable signatures; the signature for a <url,expiration_timestamp> pair depends on the time of signing and expiration.

The caching use case requires stable signatures. Will need more thought
put into this than we currently have to re-implement this feature.

for future reference:
https://blog.chef.io/cookbook-caching/
https://github.com/chef/mini_s3/blob/master/src/mini_s3.erl (expiry_time/2, etc)
https://docs.aws.amazon.com/general/latest/gr/sigv4-date-handling.html

Signed-off-by: Lincoln Baker <lbaker@chef.io>